pyproject.toml
[tool.poetry]
name = "fmp"
version = "0.1.0"
description = ""
authors = ["Carlos Fernandez <carlos.ferlo@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^2.5.3"
python-dotenv = "^1.0.0"
requests = "^2.31.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
